fix: 액션 버튼 레이아웃 변경, 헤더 영역 onPointerDownOutside 이벤트 방지
